numpy update

+++ mag2exp/ltem.py
@@ -336,7 +336,7 @@ def relativistic_wavelength(voltage):
 
     >>> import mag2exp
     >>> mag2exp.ltem.relativistic_wavelength(300e3)
-    1.9687489006848795e-12
+    np.float64(1.9687489006848795e-12)
 
     """
     return constants.h / np.sqrt(
